# src/documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "API Documentation",
      "version": "1.0.0",
      "description": "Documentation for your API"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Development server"
      }
    ],
    "tags": [
      {
        "name": "Images",
        "description": "API per la gestione delle immagini"
      }
    ],
    "paths": {
      "/api/image": {
        "post": {
          "summary": "Upload di un'immagine e ottenere informazioni sui marcatori",
          "tags": ["Images"],
          "consumes": ["multipart/form-data"],
          "parameters": [
            {
              "in": "formData",
              "name": "image",
              "type": "file",
              "required": true,
              "description": "L'immagine da caricare"
            }
          ],
          "responses": {
            "200": {
              "description": "Immagine modificata con successo",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "buffer": {
                        "type": "string",
                        "description": "L'immagine modificata"
                      },
                      "markers": {
                        "type": "array",
                        "description": "I marcatori trovati nell'immagine",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/markers": {
        "post": {
          "summary": "Ottenere informazioni sui marcatori di un'immagine",
          "tags": ["Images"],
          "consumes": ["multipart/form-data"],
          "parameters": [
            {
              "in": "formData",
              "name": "file",
              "type": "file",
              "required": true,
              "description": "L'immagine di cui ottenere i marcatori"
            }
          ],
          "responses": {
            "200": {
              "description": "Marcatori ottenuti con successo",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "markers": {
                        "type": "array",
                        "description": "I marcatori trovati nell'immagine",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/streaming": {
        "get": {
          "summary": "Avviare lo streaming di un'immagine",
          "tags": ["Images"],
          "parameters": [
            {
              "in": "query",
              "name": "stream",
              "required": true,
              "description": "URL dello stream dell'immagine",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Streaming avviato con successo"
            }
          }
        }
      }
    }
  }
  